refactor(agent): split runner into agent/runner, move orchestrator machinery out of programs
Phase 1: extract agent-runner.ts into agent/runner/:
- runner/index.ts — runAgent + runProgram dispatcher (the only isOrchestratorEnabled fork)
- runner/bootstrap.ts — bootstrapProgram + shouldDisableAsk + sessionToOptions
- runner/linear.ts — runLinearProgram (the linear arm)
- runner/types.ts — ProgramRun, BootstrapResult, AbortCase (+ re-exports)
- runner/errors.ts — abortOnInstallFailure
- agent-runner.ts reduced to a thin re-export shim

Phase 2: move programs/orchestrator/ → agent/orchestrator/:
- queue.ts, executor.ts, queue-tools.ts, run-metrics.ts and their __tests__
- orchestrator-runner.ts with updated import paths
- All importers updated: wizard-tools, agent-interface, agent-prompt-loader,
  agent-prompt-loader.test, RunScreen.tsx

No logic changes — only moves, extractions, and import fixes.
