+++ src/lib/agent/orchestrator/__tests__/executor.test.ts
@@ -5,8 +5,8 @@ import {
   QueueStore,
   type QueuedTask,
   type TaskHandoff,
-} from '@lib/programs/orchestrator/queue';
-import { drainQueue, type RunTask } from '@lib/programs/orchestrator/executor';
+} from '@lib/agent/orchestrator/queue';
+import { drainQueue, type RunTask } from '@lib/agent/orchestrator/executor';
 
 jest.mock('@utils/analytics', () => ({
   analytics: { captureException: jest.fn(), wizardCapture: jest.fn() },
