+++ src/ui/tui/screens/RunScreen.tsx
@@ -25,6 +25,7 @@ import { TipsCard } from '@ui/tui/components/TipsCard';
 import { useStdoutDimensions } from '@ui/tui/hooks/useStdoutDimensions';
 import { useFileWatcher } from '@ui/tui/hooks/file-watcher';
 import { EVENT_PLAN_FILE } from '@lib/programs/posthog-integration/index';
+
 import { getProgramConfig } from '@lib/programs/program-registry';
 import { getContentBlocks as getSkillContentBlocks } from '@lib/programs/agent-skill/content/index';
